--- backend/api/authorization/serializers.py
@@ -1,12 +1,0 @@
-from rest_framework import serializers
-
-from apps.users.models import User
-
-
-class UserLoginSerializer(serializers.ModelSerializer):
-    email = serializers.CharField()
-    password = serializers.CharField(write_only=True, required=True)
-
-    class Meta:
-        model = User
-        fields = ("email", "password")
