+++ backend/apps/authentication/urls.py
@@ -1,14 +1,23 @@
 from django.urls import path
+from rest_framework_simplejwt.views import TokenRefreshView
 
 from .views import (
+    CheckEmailView,
+    LoginView,
+    LogoutView,
     PasswordResetConfirmView,
     PasswordResetRequestView,
     RegisterView,
+    ResendVerificationView,
     VerifyEmailView,
 )
 
 urlpatterns = [
+    path("login/", LoginView.as_view(), name="login"),
+    path("refresh/", TokenRefreshView.as_view(), name="refresh-token"),
+    path("logout/", LogoutView.as_view(), name="logout"),
     path("register/", RegisterView.as_view(), name="register"),
+    path("check-email/", CheckEmailView.as_view(), name="check-email"),
     path(
         "verify/<str:uid>/<str:token>/", VerifyEmailView.as_view(), name="verify-email"
     ),
@@ -21,5 +30,10 @@ urlpatterns = [
         "password-reset/confirm/",
         PasswordResetConfirmView.as_view(),
         name="password-reset-confirm",
+    ),
+    path(
+        "resend-verification/",
+        ResendVerificationView.as_view(),
+        name="resend-verification",
     ),
 ]
